권한체크 완료

+++ src/main/java/shop/mtcoding/blog/board/BoardController.java
@@ -3,15 +3,11 @@ package shop.mtcoding.blog.board;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.AllArgsConstructor;
-import org.hibernate.internal.build.AllowPrintStacktrace;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.Mapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.*;
+import shop.mtcoding.blog._core.errors.exception.Exception401;
+import shop.mtcoding.blog._core.errors.exception.Exception403;
 import shop.mtcoding.blog.user.User;
-import shop.mtcoding.blog.user.UserRequest;
 
 import java.util.List;
 
@@ -24,19 +20,36 @@ public class BoardController {
 
     @PostMapping("/board/{id}/update")
     public String update(@PathVariable Integer id, BoardRequest.UpdateDTO reqDTO) {
-        boardRepository.updateById(id,reqDTO.getTitle(),reqDTO.getContent());
+        User sessionUser = (User) session.getAttribute("sessionUser");
+
+        Board board = boardRepository.findById(id);
+        if (sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글을 수정할 권한이 없습니다.");
+        }
+        boardRepository.updateById(id, reqDTO.getTitle(), reqDTO.getContent());
         return "redirect:/board/{id}";
     }
 
     @GetMapping("/board/{id}/update-form")
     public String updateForm(@PathVariable Integer id, HttpServletRequest req) {
         Board board = boardRepository.findById(id);
+
+        if (board == null) {
+            throw new Exception401("해당게시글을 찾을 수 없습니다.");
+        }
+
         req.setAttribute("board",board);
         return "board/update-form";
     }
 
-    @PostMapping("/board/{id}/delete")
+    @RequestMapping(value = "/board/{id}/delete", method = {RequestMethod.GET, RequestMethod.POST})
     public String delete(@PathVariable Integer id) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+
+        Board board = boardRepository.findById(id);
+        if (sessionUser.getId() != board.getUser().getId()) {
+            throw new Exception403("게시글을 삭제할 권한이 없습니다.");
+        }
         boardRepository.deleteById(id);
         return "redirect:/";
     }
@@ -61,10 +74,18 @@ public class BoardController {
         return "/board/save-form";
     }
 
+
     @GetMapping("/board/{id}")
     public String detail(@PathVariable Integer id, HttpServletRequest req) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
         Board board = boardRepository.findByIdJoinUser(id);
+
+        //로그인을 하고, 게시글의 주인이면 isOwner가 true가 된다.
+        boolean isOwenr = sessionUser.getId() == board.getUser().getId() ? true : false;
+
+        req.setAttribute("isOwenr",isOwenr);
         req.setAttribute("board",board);
+
         return "/board/detail";
     }
 }
